requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.9.1
bcrypt==4.3.0
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
crispy-bootstrap5==2025.6
cryptography==45.0.5
Django==4.2.23
django-crispy-forms==2.4
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.104.1
flake8==7.3.0
h11==0.16.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.2
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.27.0
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1

RPi.GPIO
python-dotenv
flake8
black
